fix: DB connection timing out while downloading
I haven’t confirmed this, but I got the MariaDB error 2006 along with
the mysql-error.log entry "Got timeout reading communication packets”
just before the 2006 error. I strongly suspect that this is due to
the change to using a single DB connection for much longer. Since it
leaves that connection open while downloading large video files, it
is not surprising that it would time out.