fix: source lists incorrectly being seen changed
Basically, because a list of sources of an item will likely include
sources that any single source would not include in its representation
of an item, the comparison was too general. It needed to be changed
to only look for sources in the new version of the item that are not
yet recorded in the old (or local/destination) version of the item.