fix: local Common & Mariadb.Common name collision.
I don't get why there was a problem with this since Mariadb.Common is
a submodule. But instead of wasting time digging into this, I just
renamed the local Common module to Lib which it was set to replace
anyway. Also moved Lib_conf to bin/config and renamed to Config as it
is only used in bin/ and is really specific to a specific setup.