build: add customizable app prefix
This makes it easier to accommodate IDE plugins that use .merlin
files to determine the location of libraries. For example, my setup
has APP_PREFIX_DIR set to something like "/Users/me/Projects/OVPSync"
which means that structure will be mirrored in the app container and
so .merlin files will have a path that exists on the host.