refactor: file_stores to only be paths, no usage
Tracking usage became a significant complication, so we nixed it for
now since it's not a requirement. It's much simpler to just check for
free space.

The main complication is that checking current usage is a very
expensive operation (checking every file for size) that it we wouldn't
be able to do it for each file. That meant we would do the check once
on startup and then keep track of files added and removed from each
store and update the usage number based on that. It's definitely
something we can do, but doesn't make sense at this time since we're
using storage volumes dedicated for this app and don't need to keep
space available for other purposes.

issue #4