fix: inaccurate size estimates for video files
Before the priority looked at the JW provided `size` attribute of
media to determine the size of the video files. I had forgotten that
that size is the amount of all transcodes of the video combined which
usually ends up being double to triple the size of the actual video
file we're interested in. Now we checked the "content-length" header
of each file URI to estimate the space required. If that's unavailable
we fall back to the JW provided size value.

issue #4